Updated indices for querying address history

+++ src/Lykke.Service.EthereumSamurai.MongoDb/Repositories/Erc20TransferHistoryRepository.cs
@@ -1,8 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
-using AutoMapper;
+﻿using AutoMapper;
 using Lykke.Service.EthereumSamurai.Core;
 using Lykke.Service.EthereumSamurai.Core.Repositories;
 using Lykke.Service.EthereumSamurai.Core.Settings;
@@ -10,22 +6,22 @@ using Lykke.Service.EthereumSamurai.Models.Blockchain;
 using Lykke.Service.EthereumSamurai.Models.Query;
 using Lykke.Service.EthereumSamurai.MongoDb.Entities;
 using MongoDB.Driver;
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Threading.Tasks;
 
 namespace Lykke.Service.EthereumSamurai.MongoDb.Repositories
 {
     public class Erc20TransferHistoryRepository : IErc20TransferHistoryRepository
     {
-        private readonly IBaseSettings                                _baseSettings;
-        private readonly IMapper                                      _mapper;
+        private readonly IMapper _mapper;
         private readonly IMongoCollection<Erc20TransferHistoryEntity> _historyCollection;
-
-
 
         public Erc20TransferHistoryRepository(IBaseSettings baseSettings, IMongoDatabase database, IMapper mapper)
         {
-            _baseSettings      = baseSettings;
             _historyCollection = database.GetCollection<Erc20TransferHistoryEntity>(Constants.Erc20TransferHistoryCollectionName);
-            
+
             _historyCollection.Indexes.CreateMany(new[]
             {
                 new CreateIndexModel<Erc20TransferHistoryEntity>
@@ -36,10 +32,6 @@ namespace Lykke.Service.EthereumSamurai.MongoDb.Repositories
                         Builders<Erc20TransferHistoryEntity>.IndexKeys.Ascending(x => x.TransactionIndex),
                         Builders<Erc20TransferHistoryEntity>.IndexKeys.Ascending(x => x.LogIndex)
                     )
-                ),
-                new CreateIndexModel<Erc20TransferHistoryEntity>
-                (
-                    Builders<Erc20TransferHistoryEntity>.IndexKeys.Ascending(x => x.ContractAddress)
                 ),
                 new CreateIndexModel<Erc20TransferHistoryEntity>
                 (
@@ -58,16 +50,42 @@ namespace Lykke.Service.EthereumSamurai.MongoDb.Repositories
                 (
                     Builders<Erc20TransferHistoryEntity>.IndexKeys.Combine
                     (
-                        Builders<Erc20TransferHistoryEntity>.IndexKeys.Ascending(x => x.From),
-                        Builders<Erc20TransferHistoryEntity>.IndexKeys.Ascending(x => x.ContractAddress)
+                        Builders<Erc20TransferHistoryEntity>.IndexKeys.Descending(x => x.From),
+                        Builders<Erc20TransferHistoryEntity>.IndexKeys.Descending(x => x.ContractAddress),
+                        Builders<Erc20TransferHistoryEntity>.IndexKeys.Descending(x => x.BlockNumber),
+                        Builders<Erc20TransferHistoryEntity>.IndexKeys.Ascending(x => x.TransactionIndex),
+                        Builders<Erc20TransferHistoryEntity>.IndexKeys.Ascending(x => x.LogIndex)
                     )
                 ),
                 new CreateIndexModel<Erc20TransferHistoryEntity>
                 (
                     Builders<Erc20TransferHistoryEntity>.IndexKeys.Combine
                     (
-                        Builders<Erc20TransferHistoryEntity>.IndexKeys.Ascending(x => x.To),
-                        Builders<Erc20TransferHistoryEntity>.IndexKeys.Ascending(x => x.ContractAddress)
+                        Builders<Erc20TransferHistoryEntity>.IndexKeys.Descending(x => x.From),
+                        Builders<Erc20TransferHistoryEntity>.IndexKeys.Descending(x => x.BlockNumber),
+                        Builders<Erc20TransferHistoryEntity>.IndexKeys.Ascending(x => x.TransactionIndex),
+                        Builders<Erc20TransferHistoryEntity>.IndexKeys.Ascending(x => x.LogIndex)
+                    )
+                ),
+                new CreateIndexModel<Erc20TransferHistoryEntity>
+                (
+                    Builders<Erc20TransferHistoryEntity>.IndexKeys.Combine
+                    (
+                        Builders<Erc20TransferHistoryEntity>.IndexKeys.Descending(x => x.To),
+                        Builders<Erc20TransferHistoryEntity>.IndexKeys.Descending(x => x.ContractAddress),
+                        Builders<Erc20TransferHistoryEntity>.IndexKeys.Descending(x => x.BlockNumber),
+                        Builders<Erc20TransferHistoryEntity>.IndexKeys.Ascending(x => x.TransactionIndex),
+                        Builders<Erc20TransferHistoryEntity>.IndexKeys.Ascending(x => x.LogIndex)
+                    )
+                ),
+                new CreateIndexModel<Erc20TransferHistoryEntity>
+                (
+                    Builders<Erc20TransferHistoryEntity>.IndexKeys.Combine
+                    (
+                        Builders<Erc20TransferHistoryEntity>.IndexKeys.Descending(x => x.To),
+                        Builders<Erc20TransferHistoryEntity>.IndexKeys.Descending(x => x.BlockNumber),
+                        Builders<Erc20TransferHistoryEntity>.IndexKeys.Ascending(x => x.TransactionIndex),
+                        Builders<Erc20TransferHistoryEntity>.IndexKeys.Ascending(x => x.LogIndex)
                     )
                 ),
             });
@@ -85,7 +103,7 @@ namespace Lykke.Service.EthereumSamurai.MongoDb.Repositories
         public async Task<IEnumerable<Erc20TransferHistoryModel>> GetAsync(Erc20TransferHistoryQuery query)
         {
             var filterBuilder = Builders<Erc20TransferHistoryEntity>.Filter;
-            var filter        = filterBuilder.Empty;
+            var filter = filterBuilder.Empty;
 
             if (!string.IsNullOrEmpty(query.TransactionHash))
             {
@@ -99,18 +117,35 @@ namespace Lykke.Service.EthereumSamurai.MongoDb.Repositories
 
             if (!string.IsNullOrEmpty(query.AssetHolder))
             {
-                filter &= filterBuilder.Or
-                (
-                    filterBuilder.Eq(x => x.From, query.AssetHolder),
-                    filterBuilder.Eq(x => x.To, query.AssetHolder)
-                );
+                if (string.IsNullOrEmpty(query.ContractAddress))
+                {
+                    filter &= filterBuilder.Or
+                    (
+                        filterBuilder.Eq(x => x.From, query.AssetHolder),
+                        filterBuilder.Eq(x => x.To, query.AssetHolder)
+                    );
+                }
+                else
+                {
+                    filter &= filterBuilder.Or
+                    (
+                        filterBuilder.And(
+                            filterBuilder.Eq(x => x.From, query.AssetHolder)
+                        , filterBuilder.Eq(x => x.ContractAddress, query.ContractAddress)),
+                        filterBuilder.And(filterBuilder.Eq(x => x.To, query.AssetHolder)
+                            , filterBuilder.Eq(x => x.ContractAddress, query.ContractAddress))
+                    );
+                }
             }
 
-            if (query.Contracts != null && query.Contracts.Any())
-            {
-                filter &= filterBuilder.In(x => x.ContractAddress, query.Contracts.ToList());
-            }
-            
+            var query1 = _historyCollection
+                .Find(filter)
+                .SortByDescending(x => x.BlockNumber)
+                .ThenBy(x => x.TransactionIndex)
+                .ThenBy(x => x.LogIndex)
+                .Skip(query.Start)
+                .Limit(query.Count);
+
             var entities = await _historyCollection
                 .Find(filter)
                 .SortByDescending(x => x.BlockNumber)
@@ -125,11 +160,14 @@ namespace Lykke.Service.EthereumSamurai.MongoDb.Repositories
 
         public async Task SaveForBlockAsync(IEnumerable<Erc20TransferHistoryModel> blockTransferHistory, ulong blockNumber)
         {
+            if (blockTransferHistory == null)
+                return;
+
             if (blockTransferHistory.Any(x => x.BlockNumber != blockNumber))
             {
                 throw new InvalidOperationException("All transfers should be part of the same block.");
             }
-            
+
             await _historyCollection.DeleteManyAsync(x => x.BlockNumber == blockNumber);
 
             if (blockTransferHistory.Any())
